refactor tag page design
