+++ pages/tags/[tag].tsx
@@ -2,7 +2,7 @@ import type { ReactElement } from 'react';
 import { Blog } from '../../types';
 import { Tag } from '../../data/tags/type';
 import { tags } from '../../data/tags/tags';
-import { TagContentWrapper } from '../../components/TagContentWrapper';
+import { TagContentWrapper } from '../../components/organisms/TagContentWrapper';
 import { getPostsByTags } from '../../libs/posts';
 import type { NextPageWithLayout } from '../_app';
 import { Layout } from '../../components/Layout/Layout';
