+++ tailwind.config.js
@@ -10,7 +10,6 @@ module.exports = {
       colors: {
         myBgColor: '#FFFFFF', // 全体の背景色
         myBgColorDark: '#EBEBEB', // 背景色より少し濃い色
-        myBgColorHover: '#D3BA8E',
         myTextColor: '#06283D',
         myTagColor: '#B6E388',
         myBlogListBgColor: '#F5F5F5',
